requirements.txt

# Core agent framework
llama-index 
llama-index-llms-openai 

# Web search
tavily-python



# Scraping
playwright==1.56.0
markdownify==1.2.2
pdfplumber==0.11.9
httpx==0.28.1

python-dotenv



# After installing, run this to download the Playwright browser:
# python -m playwright install chromium


